+++ processor/src/test/resources/expected/com/example/repository/Arez_CompleteRepositoryExampleRepository.java
@@ -15,7 +15,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("org.realityforge.arez.processor.ArezProcessor")
 @Singleton
-public final class Arez_CompleteRepositoryExampleRepository extends CompleteRepositoryExampleRepository implements Disposable {
+final class Arez_CompleteRepositoryExampleRepository extends CompleteRepositoryExampleRepository implements Disposable {
   private static volatile long $$arez$$_nextId;
 
   private final long $$arez$$_id;