+++ processor/src/test/resources/input/com/example/type_access_levels/ReduceAccessLevelModel.java
@@ -1,0 +1,22 @@
+package com.example.type_access_levels;
+
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Observable;
+
+@ArezComponent
+public class ReduceAccessLevelModel
+{
+  ReduceAccessLevelModel( int foo )
+  {
+  }
+
+  @Observable
+  public long getTime()
+  {
+    return 0;
+  }
+
+  public void setTime( final long time )
+  {
+  }
+}